Refactor gofaxsend into a package
Using the idiomatic Go convention of writing code in a package,
and then handle the CLI in a smaller main-package.

This makes the code importable from other parts of the code and
more maintainable.

In the spirit of portability, also split SendQfile into a separate
function that uses the Qfile reference without requiring the use
of the filesystem.