Implement in-memory variant of Qfile
As part of unifying the interface, I also made sure the "Add"
function of Qfile actually does what it describes in the doc:
creating the value if the tag does not yet exist.

Close is not part of the Qfiler interface, because (1) only the
Qfile itself needs it, and (2) because the SendQfile function
does not call it: it is only used by the SendQfileFromDisk,
function which does not use the Qfiler interface but the Qfile
directly.

Updated tests accordingly.